Crud comment + Security controllers

+++ back/controllers/user.controller.js
@@ -1,4 +1,6 @@
 const { UserModel, PostModel, CommentModel, LikeModel } = require('../models/Index');
+const token = require('../middlewares/auth.middleware');
+const { post } = require('../routes/post.routes');
 
 module.exports.getAllUsers = async (req, res) => {
 	const users = await UserModel.findAll({
@@ -17,30 +19,50 @@ module.exports.getOneUser = async (req, res) => {
 
 module.exports.updateUser = async (req, res) => {
 	// Gestion err,docs à revoir
+	const decryptedUser = token.getUserId(req);
 	try {
-		const updatedUser = await UserModel.update(
-			{ pseudo: req.body.pseudo, email: req.body.email },
-			{ where: { id: req.params.id } }
-		);
-		res.status(200).json(updatedUser);
-	} catch (err) {
-		res.status(500).send('Erreur');
+		const User = await UserModel.findOne({ where: { id: decryptedUser } });
+		if (User !== null) {
+			try {
+				const updatedUser = await UserModel.update(
+					{ pseudo: req.body.pseudo, email: req.body.email },
+					{ where: { id: req.params.id } }
+				);
+				res.status(200).json(updatedUser);
+			} catch (err) {
+				res.status(500).send('Erreur');
+			}
+		} else {
+			res.status(400).send({ message: 'Vous ne possédez pas les droits requis' });
+		}
+	} catch {
+		res.status(500).send(err);
 	}
 };
 
 module.exports.deleteUser = async (req, res) => {
+	const decryptedUser = token.getUserId(req);
 	try {
-		await UserModel.destroy({
-			where: { id: req.params.id },
-		});
+		const User = await UserModel.findOne({ where: { id: decryptedUser } });
+		if (User !== null) {
+			try {
+				await UserModel.destroy({
+					where: { id: req.params.id },
+				});
 
-		await PostModel.destroy({
-			where: { UserId: req.params.id },
-		});
-		res.cookie('jwt', '', { maxAge: 1 }); // maxAge = 1ms
-		// res.redirect('/');
-		res.status(200).json({ message: `L'utilisateur à bien été supprimé` });
-	} catch (err) {
-		return res.status(500).json({ message: err });
+				await PostModel.destroy({
+					where: { UserId: req.params.id },
+				});
+				res.cookie('jwt', '', { maxAge: 1 }); // maxAge = 1ms
+				// res.redirect('/');
+				res.status(200).json({ message: `L'utilisateur à bien été supprimé` });
+			} catch (err) {
+				return res.status(500).json({ message: err });
+			}
+		} else {
+			res.status(400).send({ message: 'Vous ne possédez pas les droits requis' });
+		}
+	} catch {
+		res.status(500).send(err);
 	}
 };
